First attempt at replacing raw NDF API with method calls
So rather than

   ndf_dim( indf )

it's now

  indf.dim()

Next step is to do the same with HDS locators and also to provide
things like "label" and "units" as actual attributes rather than
methods.

This patch probably broke module loading of starlink.ndf as I
could not quite work out how to get the Ndf.py class to load
on an "import starlink.ndf". Currently have to do

 from starlink.ndf.Ndf import Ndf

which is obviously klunky.