Add hdsNew and modify the new() method
Now if new() is called as a class method hdsNew() is called instead
of datNew() and a file will be created

  loc = hds.new( "FILE", "NAME", "TYPE )
  loc2 = loc.new( "TEST", "_DOUBLE" )

To provide a consistent user interface the new() method will now
always return a locator to the item that was created even though
the datNew() function itself will not do this.

One other important change to the new() method is that you no
longer need to provide the number of dimensions. The dimensionality
of the constructor is now an optional argument and if it is not
present the structure will be created as a scalar item. This
seems more pythonic than having to declare the dimensionality
of the numpy array itself (which numpy knows already).