Rewrite NDF mapping code
Now the map() and amap() methods return ndf.mapped objects that have
a numpytondf() method that can be used to copy from a numpy array
to a mapped NDF data array. The object tracks the number of elements
and data type internally.

Need a method for copying into a numpy array.