ndf: Tidy error handling and use NDF__MXDIM
NDF provides a constant for the max number of dimensions so we use
that instead of the NDIMX constant. We also use stack memory for
some cases where we were using malloc for things like lbnd
and ubnd.

There were some cases where raiseNdfException would not be called
to close an error context. These have now been fixed.