Import Ndf and Axis in ndfpack.
This allows the Ndf class to be imported directly from ndfpack
as shown in the documentation.  (Previously it had to be
accessed as a second level, i.e. the class was Ndf.Ndf.)

This solves the issue mentioned in the description of
commit 81e8abb76e.

Renamed the Axis and Ndf modules to lower case so they do
not clash with the names of the classes.