hds.c: split the top level functions from the locator methods
Its not clear if this is 100% the best way to do this, but it
does make the docstrings on the methods rather clearer (particularly
hds.new() to make a whole new file vs hdsloc.new to create a new
locator in an existing hds file). I don't think this should require
any changes to existing code that uses the starlink.hds module.